+++ Changelog.md
@@ -6,6 +6,7 @@
  - More browser support
  - Always run tests in build process
  - Add `uriCallback` method of getting `URI`
+ - Passing full URLs to URI.query
 
 ## Version 0.1.1